drivers: Clena serial/pl011
